refactor: rename components folder files
