+++ src/components/ui/index.ts
@@ -1,0 +1,9 @@
+export * from "./VideoPlayer";
+export * from "./VideoControls";
+export * from "./PresentationPreview";
+export * from "./Monitor";
+export * from "./Icon";
+export * from "./Carousel";
+export * from "./Button";
+
+export * from "./Slide/Slide";
